Moe 2stages (#228)
* moe_2stage_asm:add stg1_fp8_g1u1_sub128_subGU128 sp3 kernel

* clean code

* update per -token CK porting

* update : using aiter.g1u1 in CK run

* mdy name of hsa/*.co

* fix build and update ck to develop upstream

* update asm, define quantENUM

* update log

* update

* add tuner

* add more asm instance

* pass per_tensor/per_token quant

* add wint4 test

* code cleanup

* update tests

* update ck sorting fix

* rm unused moe files

* update ck

* unified Quant API

* update moe 2stage tuning

* pass all test

* add none quant case

* fix

* support moe no_quant int4_w_quant tune

* enable fp8 16x16

* enable int8 ck 2stage moe && add asm moe blockscale stage1

* update

* fix2

* update

* resolve merge issue

* fix3

* update asm co, fix blockscale

* add build process lock and fix 4

* add fmoe online tuning & fix4

* fix 5

* update

* fix

* enable tileM 32, 64, 128 only for now

* fix test moe blockscale

* update

* update

* fix issue, disable asm block gemm

* update

* reduce test number

* update

---------

+++ op_tests/test_smoothquant.py
@@ -12,7 +12,7 @@ import argparse
 @perftest()
 def run_torch(input, x_scale, y_scale_dtype=torch.float32):
     output, y_scale = aiter.pertoken_quant(
-        input, x_scale=x_scale, y_scale_dtype=y_scale_dtype)
+        input, x_scale=x_scale, scale_dtype=y_scale_dtype)
     return output, y_scale
